+++ src/containers/AudioUploader/components/AudioForm.jsx
@@ -25,7 +25,6 @@ import {
   parseCopyrightContributors,
 } from '../../../util/formHelper';
 import {
-  FormikHeader,
   FormikActionButton,
   formClasses,
   FormikAlertModalWrapper,
@@ -36,6 +35,7 @@ import { toEditAudio } from '../../../util/routeHelpers';
 import validateFormik from '../../../components/formikValidationSchema';
 import { AudioShape } from '../../../shapes';
 import * as messageActions from '../../Messages/messagesActions';
+import HeaderWithLanguage from '../../../components/SlateEditor/HeaderWithLanguage';
 
 export const getInitialValues = (audio = {}) => ({
   id: audio.id,
@@ -174,7 +174,7 @@ class AudioForm extends Component {
           const { values, dirty, isSubmitting } = formikProps;
           return (
             <Form {...formClasses()}>
-              <FormikHeader
+              <HeaderWithLanguage
                 noStatus
                 values={values}
                 type="audio"
